+++ example/ios/Runner/GeneratedPluginRegistrant.m
@@ -6,8 +6,8 @@
 
 #import "GeneratedPluginRegistrant.h"
 
-#if __has_include(<sunmi_printer_plus/SunmiPrinterPlugin.h>)
-#import <sunmi_printer_plus/SunmiPrinterPlugin.h>
+#if __has_include(<sunmi_printer_plus/SunmiPrinterPlusPlugin.h>)
+#import <sunmi_printer_plus/SunmiPrinterPlusPlugin.h>
 #else
 @import sunmi_printer_plus;
 #endif
@@ -15,7 +15,7 @@
 @implementation GeneratedPluginRegistrant
 
 + (void)registerWithRegistry:(NSObject<FlutterPluginRegistry>*)registry {
-  [SunmiPrinterPlugin registerWithRegistrar:[registry registrarForPlugin:@"SunmiPrinterPlugin"]];
+  [SunmiPrinterPlusPlugin registerWithRegistrar:[registry registrarForPlugin:@"SunmiPrinterPlusPlugin"]];
 }
 
 @end